services/commande_query/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
kafka-python
prometheus_client